content/canvas/test/test_2d.imageData.object.clamp.html

<!DOCTYPE HTML>
<title>Canvas test: 2d.imageData.object.clamp</title>
<!-- Testing: ImageData.data clamps numbers to [0, 255] -->
<script src="/MochiKit/packed.js"></script>
<script src="/tests/SimpleTest/SimpleTest.js"></script>
<link rel="stylesheet" href="/tests/SimpleTest/test.css">
<body>
<canvas id="c" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>
<script>

SimpleTest.waitForExplicitFinish();
MochiKit.DOM.addLoadEvent(function () {

var canvas = document.getElementById('c');
var ctx = canvas.getContext('2d');

var imgdata = ctx.getImageData(0, 0, 10, 10);

imgdata.data[0] = 100;
imgdata.data[0] = 300;
todo(imgdata.data[0] === 255, "imgdata.data[\""+(0)+"\"] === 255");
imgdata.data[0] = 100;
imgdata.data[0] = -100;
todo(imgdata.data[0] === 0, "imgdata.data[\""+(0)+"\"] === 0");

imgdata.data[0] = 100;
imgdata.data[0] = 200+Math.pow(2, 32);
todo(imgdata.data[0] === 255, "imgdata.data[\""+(0)+"\"] === 255");
imgdata.data[0] = 100;
imgdata.data[0] = -200-Math.pow(2, 32);
todo(imgdata.data[0] === 0, "imgdata.data[\""+(0)+"\"] === 0");

imgdata.data[0] = 100;
imgdata.data[0] = Math.pow(10, 39);
todo(imgdata.data[0] === 255, "imgdata.data[\""+(0)+"\"] === 255");
imgdata.data[0] = 100;
imgdata.data[0] = -Math.pow(10, 39);
todo(imgdata.data[0] === 0, "imgdata.data[\""+(0)+"\"] === 0");

imgdata.data[0] = 100;
imgdata.data[0] = -Infinity;
todo(imgdata.data[0] === 0, "imgdata.data[\""+(0)+"\"] === 0");
imgdata.data[0] = 100;
imgdata.data[0] = Infinity;
todo(imgdata.data[0] === 255, "imgdata.data[\""+(0)+"\"] === 255");

SimpleTest.finish();

});
</script>
